feat: move consuming background tasks to the maintenance procedure
