+++ ayon_server/addons/definition.py
@@ -6,7 +6,7 @@ import yaml
 from nxtools import log_traceback, logging, slugify
 
 from ayon_server.addons.addon import METADATA_KEYS, BaseServerAddon
-from ayon_server.addons.utils import classes_from_module, import_module
+from ayon_server.helpers.modules import classes_from_module, import_module
 
 if TYPE_CHECKING:
     from ayon_server.addons.library import AddonLibrary
