+++ maintenance/maintenance_task.py
@@ -4,22 +4,23 @@ from typing import Literal
 MaintenanceTaskType = Literal["studio", "project"]
 
 
-class MaintenanceTask(ABC):
+class BaseMaintenanceTask(ABC):
     task_type: MaintenanceTaskType
+    description: str
 
     @abstractmethod
     async def main(self, *args, **kwargs):
         pass
 
 
-class StudioMaintenanceTask(MaintenanceTask):
+class StudioMaintenanceTask(BaseMaintenanceTask):
     task_type = "studio"
 
     async def main(self):
         pass
 
 
-class ProjectMaintenanceTask(MaintenanceTask):
+class ProjectMaintenanceTask(BaseMaintenanceTask):
     task_type = "project"
 
     async def main(self, project_name: str):
